+++ src/webservices/functions.js
@@ -6,12 +6,23 @@
  */
 
 import axios from 'axios';
+import qs from 'qs';
 
-const webservice = ({ url, params }) => {
+const webservice = ({ url, params, postit = false }) => {
+  const options = {
+    method: postit ? 'POST' : 'GET',
+    headers: { 'content-type': 'application/x-www-form-urlencoded' },
+    data: qs.stringify(params),
+    url,
+  };
+
+  console.warn('webservice:OPTIONS', options);
+
+  const sender = () => (postit ? axios(options) : axios.get(url, { params }));
   return new Promise((resolve, reject) => {
-    axios
-      .get(url, { params })
+    sender()
       .then(({ data, status }) => {
+        console.warn('webservice:DATAS', data);
         if (status === 200 && data.result === 'OK') {
           resolve(data);
         } else {
@@ -19,6 +30,7 @@ const webservice = ({ url, params }) => {
         }
       })
       .catch(err => {
+        console.warn('webservice:ERROR', err);
         reject(err);
       });
   });