+++ src/webservices/index.js
@@ -5,7 +5,7 @@ export { default as webservice } from './functions';
 export { default as getIdentTour } from './getidenttour';
 export { default as getIdentVehicle } from './getidentvehicle';
 export { default as getCommands } from './getcommands';
-export { default as setGsmNumber } from './setgsmnumber';
+export { default as putGsmNumber } from './putgsmnumber';
 export { default as putSos } from './putsos';
 export { default as putWaypoint } from './putwaypoint';
 export { default as Pool } from './pool';