+++ CHANGELOG.md
@@ -1,3 +1,46 @@
+0.28.0
+
+Regarder les nouveautés des web services en ligne
+
+Finaliser l'enregistrement des données
+
+```react
+const webservice = ({ url, params, postit = false }) => {
+  const options = {
+    method: postit ? 'POST' : 'GET',
+    headers: { 'content-type': 'application/x-www-form-urlencoded' },
+    data: qs.stringify(params),
+    url,
+  };
+
+  console.warn('webservice:OPTIONS', options);
+
+  const sender = () => (postit ? axios(options) : axios.get(url, { params }));
+  return new Promise((resolve, reject) => {
+    sender()
+      .then(({ data, status }) => {
+        console.warn('webservice:DATAS', data);
+        if (status === 200 && data.result === 'OK') {
+          resolve(data);
+        } else {
+          reject((data.error && data.error) || data);
+        }
+      })
+      .catch(err => {
+        console.warn('webservice:ERROR', err);
+        reject(err);
+      });
+  });
+};
+
+```
+
+
+
+- Create installr account
+- Changer the code to put GSM Number, see `web services/putgsmnumber.js`
+- This is a good web service to use POST and GET
+
 ## 0.27.0
 
 - ADD: Merge BIP with other project